refactor: 카테고리 조회 수정
