--- backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/auth/controller/cookie/CookieExtractor.java
@@ -1,24 +1,0 @@
-package com.jjangiji.hankkimoa.auth.controller.cookie;
-
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-
-import java.util.Arrays;
-import java.util.Optional;
-
-public class CookieExtractor {
-    private static final String ACCESS_TOKEN_COOKIE_NAME = "accessToken";
-
-    public static Optional<String> extractAccessToken(HttpServletRequest request) {
-        Cookie[] cookies = request.getCookies();
-
-        if (cookies == null) {
-            return Optional.empty();
-        }
-
-        return Arrays.stream(cookies)
-                .filter(cookie -> ACCESS_TOKEN_COOKIE_NAME.equals(cookie.getName()))
-                .map(Cookie::getValue)
-                .findFirst();
-    }
-}
