+++ backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/auth/controller/cookie/CookieResolver.java
@@ -63,4 +63,3 @@ public class CookieResolver {
         return isAccessTokenEmpty(request) && isRefreshTokenEmpty(request);
     }
 }
-
