--- backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/user/UserService.java
@@ -1,22 +1,0 @@
-package com.jjangiji.hankkimoa.user;
-
-import com.jjangiji.hankkimoa.common.exception.ExceptionCode;
-import com.jjangiji.hankkimoa.common.exception.HankkiMoaException;
-import com.jjangiji.hankkimoa.user.domain.User;
-import com.jjangiji.hankkimoa.user.repository.UserRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@RequiredArgsConstructor
-@Service
-public class UserService {
-    private final UserRepository userRepository;
-
-    @Transactional(readOnly = true)
-    public UserMeResponse getMyInfo(Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new HankkiMoaException(ExceptionCode.USER_NOT_FOUND));
-        return UserMeResponse.from(user);
-    }
-}
