+++ backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/user/repository/UserCategoryRepository.java
@@ -1,7 +1,12 @@
 package com.jjangiji.hankkimoa.user.repository;
 
+import com.jjangiji.hankkimoa.user.domain.User;
 import com.jjangiji.hankkimoa.user.domain.UserCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
+import java.util.Optional;
+
 public interface UserCategoryRepository extends JpaRepository<UserCategory, Long> {
+    List<UserCategory> findAllByUser(User user);
 }
